Refactoring to encapsulate categories and make it easier for new devs.
So now the idea is that categories will be self contained in a dir called
categories at the top level. Templates and static resources for that category
will all live in that dir.

Maybe we should rename categories (and our url param category) to test_sets?

All the app-fundamentals code is going to live in base/ and new devs
shouldn't have to edit any of that.

Also fixed the 2 hanging unit tests in test_categories.html